rename avm2 directory to abc
