+++ pyfdec/abc/ConstantKind.py
