+++ pyfdec/tags/DoABC2.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import ClassVar
 from pyfdec.extended_buffer import ExtendedBuffer
 from pyfdec.tags.Tag import Tag
-from pyfdec.avm2.ABCFile import ABCFile
+from pyfdec.abc.ABCFile import ABCFile
 
 
 @dataclass
